[net tutorial]: remove inconsequntial diff content

+++ examples/tutorials/thread_network/08_screen_u8g2/main.c
@@ -16,7 +16,7 @@ static void update_screen(void);
 
 uint8_t global_temperature_setpoint = 0;
 uint8_t local_temperature_setpoint  = 22;
-uint8_t measured_temperature        = 0;
+int measured_temperature = 0;
 
 static void button_callback(returncode_t ret,
                             int          btn_num,
